implement threading; add swagger

+++ src/main/java/com/prep/customer/purchasing/repository/TransactionRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface TransactionRepository extends JpaRepository<Transaction, Integer> {
-    public List<Transaction> findByCustomerId(Integer custId);
+    public List<Transaction> findByCustomerId(Long custId);
 }
